// applications/webapp/src/app/api/generated/model/run-item.ts
/**
 * VidIngest API
 *
 * Contact: dev@tradinglabs.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RunItem { 
    itemId?: string;
    url?: string;
    status?: string;
    phase?: string;
    failedPhase?: string;
    phaseUpdatedAt?: string;
    errorCode?: string;
    error?: string;
    videoId?: string;
    channelName?: string;
    videoTitle?: string;
    attempt?: number;
}